added routing to the app + implemented tab bar view
